+++ spec/compile_spec.lua
@@ -100,6 +100,12 @@ describe('arrow function #5.1+', function()
       local a = x -> { return x }
       return a(1)
     ]])
+    assert.run(1, [[
+      local a = x -> { return x }
+      return a(1)
+    ]])
+    local sum = runErde('return (a, b) -> a + b')
+    assert.are.equal(3, sum(1, 2))
   end)
 
   spec('fat', function()
@@ -114,6 +120,8 @@ describe('arrow function #5.1+', function()
       a.c = x => { return self.b + x }
       return a:c(2)
     ]])
+    local sum = runErde('return (a, b) => a + b + self.c')
+    assert.are.equal(6, sum({ c = 3 }, 1, 2))
   end)
 
   spec('iife', function()